@@ -1,4 +1,4 @@
-Sanataro::Application.routes.draw do
+Rails.application.routes.draw do
   use_doorkeeper
 
   root to: "login#login"
