@@ -4,15 +4,8 @@ module ArelExt
   module Visitors
     module PostgreSQL
       module InstanceMethods
-        def self.included(base)
-          base.class_eval do
-            alias_method_chain :visit_Arel_Nodes_Matches, :format_sql92
-            alias_method_chain :visit_Arel_Nodes_DoesNotMatch, :format_sql92
-          end
-        end
-
-        def visit_Arel_Nodes_Matches_with_format_sql92 o, a
-          collector = visit_Arel_Nodes_Matches_without_format_sql92(o, a)
+        def visit_Arel_Nodes_Matches o, a
+          collector = super
           if o.escape
             collector << ' ESCAPE '
             visit o.escape, collector
@@ -21,8 +14,8 @@ module ArelExt
           end
         end
 
-        def visit_Arel_Nodes_DoesNotMatch_with_format_sql92 o, a
-          collector = visit_Arel_Nodes_DoesNotMatch_without_format_sql92(o, a)
+        def visit_Arel_Nodes_DoesNotMatch o, a
+          collector = super
           if o.escape
             collector << ' ESCAPE '
             visit o.escape, collector
@@ -36,5 +29,5 @@ module ArelExt
 end
 
 ActiveSupport.on_load(:active_record) do
-  Arel::Visitors::PostgreSQL.send(:include, ArelExt::Visitors::PostgreSQL::InstanceMethods)
+  Arel::Visitors::PostgreSQL.prepend ArelExt::Visitors::PostgreSQL::InstanceMethods
 end
