@@ -57,7 +57,7 @@ describe ChartData::AssetsController, type: :controller do
           get :show, id: '200301', format: :json
         end
         subject { response }
-        it { is_expected.to be_success }
+        it { is_expected.to be_successful }
 
         describe '#body' do
           subject { super().body }
@@ -88,7 +88,7 @@ describe ChartData::AssetsController, type: :controller do
 
           describe 'response' do
             subject { response }
-            it { is_expected.to be_success }
+            it { is_expected.to be_successful }
             specify do
               expect(ActiveSupport::JSON.decode(subject.body)).to eq([{ 'label' => 'その1', 'data' => 800 }, { 'label' => 'その3', 'data' => 900 }])
             end
@@ -101,7 +101,7 @@ describe ChartData::AssetsController, type: :controller do
 
           describe 'response' do
             subject { response }
-            it { is_expected.to be_success }
+            it { is_expected.to be_successful }
             specify do
               expect(ActiveSupport::JSON.decode(subject.body)).to eq([{ 'label' => 'その4', 'data' => 200 }])
             end
