@@ -31,7 +31,7 @@ describe ConfirmationRequiredsController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template 'update' }
         end
 
@@ -59,7 +59,7 @@ describe ConfirmationRequiredsController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template 'update' }
         end
 
@@ -92,7 +92,7 @@ describe ConfirmationRequiredsController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template 'update' }
         end
 
@@ -120,7 +120,7 @@ describe ConfirmationRequiredsController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template 'update' }
         end
       end
@@ -138,7 +138,7 @@ describe ConfirmationRequiredsController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template 'update' }
         end
       end
