@@ -25,7 +25,7 @@ describe ProfitLossesController, type: :controller do
 
         describe 'response' do
           subject { response }
-          it { is_expected.to be_success }
+          it { is_expected.to be_successful }
           it { is_expected.to render_template('index') }
         end
 
@@ -77,7 +77,7 @@ describe ProfitLossesController, type: :controller do
           end
           describe 'response' do
             subject { response }
-            it { is_expected.to be_success }
+            it { is_expected.to be_successful }
             it { is_expected.to render_template('index') }
           end
 
@@ -203,7 +203,7 @@ describe ProfitLossesController, type: :controller do
 
           describe 'response' do
             subject { response }
-            it { is_expected.to be_success }
+            it { is_expected.to be_successful }
             it { is_expected.to render_template 'show' }
           end
           describe 'assigned variables' do
