@@ -27,7 +27,7 @@ describe Settings::UsersController, type: :controller do
         end
 
         subject { response }
-        it { is_expected.to be_success }
+        it { is_expected.to be_successful }
         it { is_expected.to render_template 'show' }
       end
     end
@@ -66,7 +66,7 @@ describe Settings::UsersController, type: :controller do
 
           describe 'response' do
             subject { response }
-            it { is_expected.to be_success }
+            it { is_expected.to be_successful }
             it { is_expected.to render_template 'update' }
           end
 
