@@ -25,7 +25,7 @@ describe TagStatusesController, type: :controller do
 
       describe 'response' do
         subject { response }
-        it { is_expected.to be_success }
+        it { is_expected.to be_successful }
         it { is_expected.to render_template 'show' }
       end
 
